requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2==2.9.9
python-decouple==3.8
requests==2.31.0

